[DDW-683] Refactor the Loading screen (#1446)
* [DDW-683] Files reorganization

* [DDW-683] CHANGELOG

* [DDW-683] Data adjustments

* [DDW-683] Remove old files

* [DDW-683] Update stories

* [DDW-683] Prettier and Translations manager

* [DDW-683] Separated Connecting Syncing into smaller components

* [DDW-683] Translations manager

* [DDW-683] Organizing methods

* Merge branch 'develop' into feature/ddw-683-refactor-the-loading-screen

* Merge branch 'develop' into feature/ddw-683-refactor-the-loading-screen

* [DDW-683] Fix failing acceptance tests, Imporove component namings

* [DDW-683] Improve storybook, Fix loading screen notifications styling

* [DDW-683] Storybook cleanup

* [DDW-683] External link icons improvement

* [DDW-683] Restore overlays
