--- source/renderer/app/components/loading/Loading.scss
@@ -1,174 +1,0 @@
-@import '../../themes/mixins/animations';
-
-.component {
-  height: 100%;
-  transition: background-color 500ms;
-  width: 100%;
-
-  .reportIssue {
-    position: fixed;
-    text-align: center;
-    width: 100%;
-
-    .reportIssueText {
-      font-family: var(--font-regular);
-      font-size: 18px;
-      line-height: 1.22;
-      margin-top: 24px;
-    }
-
-    .actionButton {
-      background-color: var(--theme-report-issue-button-background-color);
-      margin: 20px 10px 10px;
-      width: 240px;
-
-      &:hover {
-        background-color: var(
-          --theme-report-issue-button-background-color-hover
-        ) !important;
-      }
-      &:active {
-        background-color: var(
-          --theme-report-issue-button-background-color-active
-        ) !important;
-      }
-    }
-
-    .downloadLogsButton {
-      border-bottom: 1px solid
-        var(--theme-report-issue-syncing-download-logs-text-color);
-      color: var(--theme-report-issue-syncing-download-logs-text-color);
-      cursor: pointer;
-      font-family: var(--font-regular);
-      font-size: 14px;
-      line-height: 1.36;
-      margin-bottom: 20px;
-      opacity: 0.8;
-      &:disabled {
-        color: var(--rp-button-bg-color-disabled);
-        text-decoration: none;
-        cursor: default;
-        border-bottom: none;
-      }
-    }
-    .downloadLogsButtonConnecting {
-      border-bottom: 1px solid
-        var(--theme-report-issue-syncing-download-logs-text-color-connecting);
-      color: var(
-        --theme-report-issue-syncing-download-logs-text-color-connecting
-      );
-    }
-  }
-  &.is-loading-theme {
-    background-color: transparent !important;
-  }
-  &.is-connecting {
-    background-color: var(--theme-connecting-background-color);
-
-    .reportIssue {
-      background-color: var(--theme-report-issue-connecting-background-color);
-    }
-    .reportIssueText {
-      color: var(--theme-report-issue-connecting-text-color);
-    }
-  }
-  &.is-syncing {
-    background-color: var(--theme-syncing-background-color);
-
-    .reportIssue {
-      background-color: var(--theme-report-issue-syncing-background-color);
-    }
-    .reportIssueText {
-      color: var(--theme-report-issue-syncing-text-color);
-    }
-  }
-}
-
-.logos {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-}
-
-.connecting,
-.syncing {
-  bottom: 64px;
-  position: absolute;
-  text-align: center;
-  width: 100%;
-}
-
-.syncing {
-  color: var(--theme-syncing-text-color);
-}
-
-.ada-logo,
-.ada-apiLogo {
-  & > svg {
-    display: block;
-    width: 48px;
-    opacity: 0.5;
-  }
-}
-
-$cardanoLogoWidth: 86px;
-.ada-apiLogo > svg {
-  width: $cardanoLogoWidth;
-}
-
-$adaLogoWidth: 43px;
-.ada-logo > svg {
-  width: $adaLogoWidth;
-  margin: 3px ($cardanoLogoWidth - $adaLogoWidth) / 2 0; // Visually align in the middle + Needs to be the same dimension as the apiLogo
-}
-
-.daedalusLogo {
-  & > svg {
-    display: block;
-    width: 173px;
-    margin: 0 114px;
-  }
-}
-
-.connectingLogo {
-  & > svg path {
-    fill: var(--theme-icon-connecting-logo-color);
-  }
-}
-
-.syncingLogo svg path {
-  fill: var(--theme-icon-syncing-logo-color);
-}
-
-.linkNewWindow {
-  display: inline-block;
-  margin-right: 10px;
-  vertical-align: middle;
-  width: 12px;
-  g {
-    stroke: var(--theme-report-issue-icon-color);
-  }
-}
-
-.headline {
-  font-family: var(--font-regular);
-  font-size: 18px;
-}
-
-.connecting {
-  color: var(--theme-connecting-text-color);
-  .headline {
-    overflow: visible;
-    &:not(.withoutAnimation) {
-      @include animated-ellipsis($width: 16px);
-    }
-  }
-}
-
-.component,
-.headline,
-.logo {
-  -webkit-backface-visibility: hidden;
-}
