+++ source/renderer/app/components/loading/manual-update/ManualUpdate.js
@@ -6,35 +6,35 @@ import { defineMessages, intlShape, FormattedHTMLMessage } from 'react-intl';
 import ReactModal from 'react-modal';
 import { Button } from 'react-polymorph/lib/components/Button';
 import { ButtonSkin } from 'react-polymorph/lib/skins/simple/ButtonSkin';
-import linkNewWindow from '../../assets/images/link-ic.inline.svg';
-import styles from './ManualUpdateOverlay.scss';
+import linkNewWindow from '../../../assets/images/link-ic.inline.svg';
+import styles from './ManualUpdate.scss';
 
 const messages = defineMessages({
   title: {
-    id: 'manualUpdateOverlay.title',
+    id: 'manualUpdate.title',
     defaultMessage: '!!!Software update is available',
     description: 'Title for "Manual update" overlay',
   },
   descriptionLine1: {
-    id: 'manualUpdateOverlay.description1',
+    id: 'manualUpdate.description1',
     defaultMessage:
       '!!!You are experiencing network connection issues, and you are not running the latest Daedalus version. Automatic updates are unavailable while Daedalus is not connected to Cardano network.',
     description: 'Description line 1 of "Manual update" overlay',
   },
   descriptionLine2: {
-    id: 'manualUpdateOverlay.description2',
+    id: 'manualUpdate.description2',
     defaultMessage:
       '!!!You are currently running <b>{currentAppVersion}</b> version of Daedalus, and <b>{availableAppVersion}</b> version is available. Please manually update to that version since it may resolve your connecting issues.',
     description: 'Description line 2 of "Manual update" overlay',
   },
   actionButtonLabel: {
-    id: 'manualUpdateOverlay.button.label',
+    id: 'manualUpdate.button.label',
     defaultMessage: '!!!Follow instructions and manually update',
     description:
       'Label for "Follow instructions and manually update" action button on "Manual update" overlay',
   },
   manualUpdateButtonUrl: {
-    id: 'manualUpdateOverlay.button.url',
+    id: 'manualUpdate.button.url',
     defaultMessage:
       '!!!https://iohk.zendesk.com/hc/en-us/articles/360023850634',
     description:
@@ -49,7 +49,7 @@ type Props = {
 };
 
 @observer
-export default class ManualUpdateOverlay extends Component<Props> {
+export default class ManualUpdate extends Component<Props> {
   static contextTypes = {
     intl: intlShape.isRequired,
   };
