+++ source/renderer/app/components/loading/manual-update/ManualUpdate.scss
@@ -1,4 +1,4 @@
-@import '../../themes/mixins/link';
+@import '../../../themes/mixins/link';
 
 .overlay {
   bottom: 0;
