+++ source/renderer/app/components/loading/no-disk-space-error/NoDiskSpaceError.scss
@@ -1,4 +1,4 @@
-@import '../../themes/mixins/animations';
+@import '../../../themes/mixins/animations';
 
 .component {
   align-items: center;
