+++ source/renderer/app/components/loading/syncing-connecting/ReportIssue.scss
@@ -1,0 +1,81 @@
+@import '../../../themes/mixins/link';
+
+.component {
+  position: fixed;
+  text-align: center;
+  width: 100%;
+
+  &.is-connecting {
+    background-color: var(--theme-report-issue-connecting-background-color);
+    .reportIssueText {
+      color: var(--theme-report-issue-connecting-text-color);
+    }
+  }
+  &.is-syncing {
+    background-color: var(--theme-report-issue-syncing-background-color);
+    .reportIssueText {
+      color: var(--theme-report-issue-syncing-text-color);
+    }
+  }
+}
+
+.reportIssueText {
+  font-family: var(--font-regular);
+  font-size: 18px;
+  line-height: 1.22;
+  margin-top: 24px;
+}
+
+.actionButton {
+  background-color: var(--theme-report-issue-button-background-color);
+  margin: 20px 10px 10px;
+  width: 240px;
+
+  &:hover {
+    background-color: var(
+      --theme-report-issue-button-background-color-hover
+    ) !important;
+  }
+  &:active {
+    background-color: var(
+      --theme-report-issue-button-background-color-active
+    ) !important;
+  }
+
+  .externalLinkIcon {
+    @include link(--theme-button-primary-text-color);
+    border-bottom: none;
+    margin-right: 12px;
+    svg {
+      margin-left: 0;
+      vertical-align: sub;
+      width: 15px;
+      g {
+        stroke-width: 3;
+      }
+    }
+  }
+}
+
+.downloadLogsButton {
+  border-bottom: 1px solid
+    var(--theme-report-issue-syncing-download-logs-text-color);
+  color: var(--theme-report-issue-syncing-download-logs-text-color);
+  cursor: pointer;
+  font-family: var(--font-regular);
+  font-size: 14px;
+  line-height: 1.36;
+  margin-bottom: 20px;
+  opacity: 0.8;
+  &:disabled {
+    border-bottom-color: transparent;
+    cursor: default;
+    text-decoration: none;
+  }
+}
+
+.downloadLogsButtonConnecting {
+  border-bottom: 1px solid
+    var(--theme-report-issue-syncing-download-logs-text-color-connecting);
+  color: var(--theme-report-issue-syncing-download-logs-text-color-connecting);
+}
