+++ source/renderer/app/components/loading/syncing-connecting/StatusIcons-tooltip.scss
