+++ source/renderer/app/components/loading/syncing-connecting/StatusIcons.js
@@ -7,13 +7,13 @@ import { TooltipSkin } from 'react-polymorph/lib/skins/simple/TooltipSkin';
 import classNames from 'classnames';
 import styles from './StatusIcons.scss';
 import tooltipStyles from './StatusIcons-tooltip.scss';
-import { CardanoNodeStates } from '../../../../common/types/cardano-node.types';
-import nodeStateIcon from '../../assets/images/node-state-icon.inline.svg';
-import isNodeRespondingIcon from '../../assets/images/is-node-responding-icon.inline.svg';
-import isNodeSubscribedIcon from '../../assets/images/is-node-subscribed-icon.inline.svg';
-import isNodeTimeCorrectIcon from '../../assets/images/is-node-time-correct-icon.inline.svg';
-import isNodeSyncingIcon from '../../assets/images/is-node-syncing-icon.inline.svg';
-import type { CardanoNodeState } from '../../../../common/types/cardano-node.types';
+import { CardanoNodeStates } from '../../../../../common/types/cardano-node.types';
+import nodeStateIcon from '../../../assets/images/node-state-icon.inline.svg';
+import isNodeRespondingIcon from '../../../assets/images/is-node-responding-icon.inline.svg';
+import isNodeSubscribedIcon from '../../../assets/images/is-node-subscribed-icon.inline.svg';
+import isNodeTimeCorrectIcon from '../../../assets/images/is-node-time-correct-icon.inline.svg';
+import isNodeSyncingIcon from '../../../assets/images/is-node-syncing-icon.inline.svg';
+import type { CardanoNodeState } from '../../../../../common/types/cardano-node.types';
 
 const messages = defineMessages({
   nodeIsRunning: {
