+++ source/renderer/app/components/loading/syncing-connecting/StatusIcons.scss
@@ -1,4 +1,4 @@
-@import '../../themes/mixins/animations';
+@import '../../../themes/mixins/animations';
 
 .component {
   bottom: 20px;
