+++ source/renderer/app/components/loading/syncing-connecting/SyncingConnecting.scss
@@ -1,0 +1,20 @@
+.component {
+  height: 100%;
+  transition: background-color 500ms;
+  width: 100%;
+
+  &.is-loading-theme {
+    background-color: transparent !important;
+  }
+  &.is-connecting {
+    background-color: var(--theme-connecting-background-color);
+  }
+  &.is-syncing {
+    background-color: var(--theme-syncing-background-color);
+  }
+}
+
+.component,
+.logo {
+  -webkit-backface-visibility: hidden;
+}
