+++ source/renderer/app/components/loading/system-time-error/SystemTimeError.scss
@@ -1,5 +1,5 @@
-@import '../../themes/mixins/animations';
-@import '../../themes/mixins/link';
+@import '../../../themes/mixins/animations';
+@import '../../../themes/mixins/link';
 
 .component {
   align-items: center;
