+++ source/renderer/app/containers/Root.js
@@ -2,7 +2,7 @@
 import React, { Component } from 'react';
 import { inject, observer } from 'mobx-react';
 import WalletAddPage from './wallet/WalletAddPage';
-import LoadingPage from './LoadingPage';
+import Loading from './loading/Loading';
 import { DIALOGS } from '../../../common/ipc/constants';
 import type { InjectedContainerProps } from '../types/injectedPropsType';
 
@@ -59,7 +59,7 @@ export default class Root extends Component<Props> {
       !isSystemTimeCorrect ||
       isNotEnoughDiskSpace
     ) {
-      return <LoadingPage stores={stores} actions={actions} />;
+      return <Loading stores={stores} actions={actions} />;
     }
 
     if (!wallets.hasAnyWallets) {
