+++ source/renderer/app/i18n/locales/en-US.json
@@ -152,11 +152,11 @@
   "loading.screen.unrecoverableCardanoMessage": "Unable to start Cardano node. Please submit a support request.",
   "loading.screen.updatedCardanoMessage": "Cardano node updated",
   "loading.screen.updatingCardanoMessage": "Updating Cardano node",
-  "manualUpdateOverlay.button.label": "Follow instructions and manually update",
-  "manualUpdateOverlay.button.url": "https://iohk.zendesk.com/hc/en-us/articles/360023850634",
-  "manualUpdateOverlay.description1": "You are experiencing network connection issues, and you are not running the latest Daedalus version. Automatic updates are unavailable while Daedalus is not connected to Cardano network.",
-  "manualUpdateOverlay.description2": "You are currently running <b>{currentAppVersion}</b> version of Daedalus, and <b>{availableAppVersion}</b> version is available. Please manually update to that version since it may resolve your connecting issues.",
-  "manualUpdateOverlay.title": "Software update is available",
+  "manualUpdate.button.label": "Follow instructions and manually update",
+  "manualUpdate.button.url": "https://iohk.zendesk.com/hc/en-us/articles/360023850634",
+  "manualUpdate.description1": "You are experiencing network connection issues, and you are not running the latest Daedalus version. Automatic updates are unavailable while Daedalus is not connected to Cardano network.",
+  "manualUpdate.description2": "You are currently running <b>{currentAppVersion}</b> version of Daedalus, and <b>{availableAppVersion}</b> version is available. Please manually update to that version since it may resolve your connecting issues.",
+  "manualUpdate.title": "Software update is available",
   "noDiskSpace.error.overlayContent": "<b>Daedalus requires at least {diskSpaceRequired} of hard drive space to operate. Your computer is missing {diskSpaceMissing} of available space. Please delete some files to increase available hard drive space to continue using Daedalus. </b><br /><br />It is recommended to have at least 15% of hard drive space available ({diskSpaceRecommended} in your case) for normal and stable operation of the operating system and installed programs. We strongly recommend that you free up at least that amount of space from your hard drive.",
   "noDiskSpace.error.overlayTitle": "Daedalus requires more hard drive space",
   "paper.wallet.create.certificate.completion.dialog.addressCopiedLabel": "copied",
