+++ source/renderer/app/i18n/locales/ja-JP.json
@@ -152,11 +152,11 @@
   "loading.screen.unrecoverableCardanoMessage": "カルダノ・ノードの起動に失敗しました。サポートセンターに問い合わせてください。",
   "loading.screen.updatedCardanoMessage": "カルダノ・ノードがアップデートされました",
   "loading.screen.updatingCardanoMessage": "カルダノ・ノードをアップデート中",
-  "manualUpdateOverlay.button.label": "指示に従って手動で更新してください",
-  "manualUpdateOverlay.button.url": "https://iohk.zendesk.com/hc/ja/articles/360023850634",
-  "manualUpdateOverlay.description1": "ネットワークへの接続に不具合があり、Daedalusウォレットの最新バージョンが実行されていません。DaedalusがCardanoネットワークに接続されていない間、自動更新は実行できません。",
-  "manualUpdateOverlay.description2": "ご使用のDaedalusはバージョン<b>{currentAppVersion}</b>です。現在<b>{availableAppVersion}</b>をご利用いただけます。手動で最新版に更新すると、接続の不具合を解消できる場合があります。",
-  "manualUpdateOverlay.title": "ソフトウェアが更新できます",
+  "manualUpdate.button.label": "指示に従って手動で更新してください",
+  "manualUpdate.button.url": "https://iohk.zendesk.com/hc/ja/articles/360023850634",
+  "manualUpdate.description1": "ネットワークへの接続に不具合があり、Daedalusウォレットの最新バージョンが実行されていません。DaedalusがCardanoネットワークに接続されていない間、自動更新は実行できません。",
+  "manualUpdate.description2": "ご使用のDaedalusはバージョン<b>{currentAppVersion}</b>です。現在<b>{availableAppVersion}</b>をご利用いただけます。手動で最新版に更新すると、接続の不具合を解消できる場合があります。",
+  "manualUpdate.title": "ソフトウェアが更新できます",
   "noDiskSpace.error.overlayContent": "<b>ダイダロスを動作させるには、ハードディスクに少なくとも{diskSpaceRequired}の空き容量が必要です。お使いのコンピュータは、空き容量が{diskSpaceMissing}不足しています。ダイダロスのご利用を続けるためには、ファイルをいくつか削除して空き容量を増やしてください。 </b><br /><br />オペレーティングシステムとインストール済みプログラムを正常かつ安定した状態で動作させるには、ハードディスクに少なくとも15％（お使いのコンピュータの場合は{diskSpaceRecommended}）の空き容量が必要です。ハードディスクに少なくともこの程度の空き容量を確保することを強くお勧めします。",
   "noDiskSpace.error.overlayTitle": "ダイダロスを動作させるには、ハードディスクの空き容量が足りません",
   "paper.wallet.create.certificate.completion.dialog.addressCopiedLabel": "コピーされました",
