+++ source/renderer/app/i18n/locales/zh-CN.json
@@ -152,11 +152,11 @@
   "loading.screen.unrecoverableCardanoMessage": "!!!Unable to start Cardano node. Please submit a support request.",
   "loading.screen.updatedCardanoMessage": "!!!Cardano node updated",
   "loading.screen.updatingCardanoMessage": "!!!Updating Cardano node",
-  "manualUpdateOverlay.button.label": "!!!Follow instructions and manually update",
-  "manualUpdateOverlay.button.url": "!!!https://iohk.zendesk.com/hc/en-us",
-  "manualUpdateOverlay.description1": "!!!You are experiencing network connection issues, and you are not running the latest Daedalus version. Automatic updates are unavailable while Daedalus is not connected to Cardano network.",
-  "manualUpdateOverlay.description2": "!!!You are currently running <b>{currentAppVersion}</b> version of Daedalus, and <b>{availableAppVersion}</b> version is available. Please manually update to that version since it may resolve your connecting issues.",
-  "manualUpdateOverlay.title": "!!!Software update is available",
+  "manualUpdate.button.label": "!!!Follow instructions and manually update",
+  "manualUpdate.button.url": "!!!https://iohk.zendesk.com/hc/en-us",
+  "manualUpdate.description1": "!!!You are experiencing network connection issues, and you are not running the latest Daedalus version. Automatic updates are unavailable while Daedalus is not connected to Cardano network.",
+  "manualUpdate.description2": "!!!You are currently running <b>{currentAppVersion}</b> version of Daedalus, and <b>{availableAppVersion}</b> version is available. Please manually update to that version since it may resolve your connecting issues.",
+  "manualUpdate.title": "!!!Software update is available",
   "noDiskSpace.error.overlayContent": "!!!<b>There is not enough disk space left on your device.</b><br />Daedalus requires at least {diskSpaceRequired} to operate. Please free up some disk space to continue.",
   "noDiskSpace.error.overlayTitle": "!!!Not enough disk space",
   "paper.wallet.create.certificate.completion.dialog.addressCopiedLabel": "!!!copied",
