+++ source/renderer/app/themes/daedalus/cardano.js
@@ -572,7 +572,7 @@ const CARDANO_THEME_CONFIG = {
     '--theme-report-issue-icon-color': '#fafbfc',
     '--theme-report-issue-syncing-background-color': 'rgba(94, 96, 102, 0.05)',
     '--theme-report-issue-syncing-text-color': '#5e6066',
-    '--theme-report-issue-syncing-download-logs-text-color': '#fafbfc',
+    '--theme-report-issue-syncing-download-logs-text-color': '#5e6066',
     '--theme-report-issue-syncing-download-logs-text-color-connecting':
       '#fafbfc',
   },
