+++ storybook/stories/Loading.stories.js
@@ -1,24 +1,20 @@
 // @flow
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import {
-  withKnobs,
-  number,
-  boolean,
-  text,
-  select,
-} from '@storybook/addon-knobs';
-import { linkTo } from '@storybook/addon-links';
-import { action } from '@storybook/addon-actions';
+import { withKnobs } from '@storybook/addon-knobs';
 
 // Assets and helpers
 import StoryDecorator from './support/StoryDecorator';
 
 // Screens
-import Loading from '../../source/renderer/app/components/loading/Loading';
-import currencyIcon from '../../source/renderer/app/assets/images/ada-logo.inline.svg';
-import apiIcon from '../../source/renderer/app/assets/images/cardano-logo.inline.svg';
-import { CardanoNodeStates } from '../../source/common/types/cardano-node.types';
+import {
+  DefaultSyncingConnectingStory,
+  ConnectivityIssuesSyncingConnectingStory,
+  SyncIssuesSyncingConnectingStory,
+} from './Loading-SyncingConnecting.stories';
+import { NoDiskSpaceErrorStory } from './Loading-NoDiskSpaceError.stories';
+import { SystemTimeErrorStory } from './Loading-SystemTimeError.stories';
+import { ManualUpdateStory } from './Loading-ManualUpdate.stories';
 
 storiesOf('Loading', module)
   .addDecorator((story, context) => (
@@ -27,165 +23,12 @@ storiesOf('Loading', module)
 
   // ====== Stories ======
 
-  .add('Default', () => (
-    <Loading
-      onStatusIconClick={linkTo('Diagnostics', () => 'default')}
-      currencyIcon={currencyIcon}
-      apiIcon={apiIcon}
-      cardanoNodeState={select(
-        'cardanoNodeState',
-        CardanoNodeStates,
-        CardanoNodeStates.STARTING
-      )}
-      availableAppVersion={null}
-      hasBeenConnected={boolean('hasBeenConnected', true)}
-      isConnected={boolean('isConnected', true)}
-      isSynced={boolean('isSynced', true)}
-      isNodeStopping={boolean('isNodeStopping', false)}
-      isNodeStopped={boolean('isNodeStopped', false)}
-      isNotEnoughDiskSpace={boolean('isNotEnoughDiskSpace', false)}
-      isTlsCertInvalid={boolean('isTlsCertInvalid', false)}
-      diskSpaceRequired={`${text('diskSpaceRequired (GB)', 4)} GB`}
-      diskSpaceMissing={`${text('diskSpaceRequired (GB)', 1)} GB`}
-      diskSpaceRecommended={`${text('diskSpaceRequired (GB)', 8)} GB`}
-      syncPercentage={number('syncPercentage', 100)}
-      loadingDataForNextScreenMessage={{
-        id: 'loading.screen.loadingWalletData',
-        defaultMessage: '!!!Loading wallet data',
-        description: 'Message "Loading wallet data" on the loading screen.',
-      }}
-      hasLoadedCurrentLocale={boolean('hasLoadedCurrentLocale', true)}
-      hasLoadedCurrentTheme={boolean('hasLoadedCurrentTheme', true)}
-      localTimeDifference={number('localTimeDifference', 0)}
-      isSystemTimeCorrect={boolean('isSystemTimeCorrect', true)}
-      isCheckingSystemTime={boolean('isCheckingSystemTime', false)}
-      isNodeResponding={boolean('isNodeResponding', true)}
-      isNodeSubscribed={boolean('isNodeSubscribed', true)}
-      isNodeSyncing={boolean('isNodeSyncing', true)}
-      isNodeTimeCorrect={boolean('isNodeTimeCorrect', true)}
-      currentLocale="en-US"
-      currentAppVersion="1.1.1"
-      isNewAppVersionAvailable={boolean('isNewAppVersionAvailable', false)}
-      isNewAppVersionLoading={boolean('isNewAppVersionLoading', false)}
-      isNewAppVersionLoaded={boolean('isNewAppVersionLoaded', false)}
-      onExternalLinkClick={action('onExternalLinkClick')}
-      onReportIssueClick={action('onReportIssueClick')}
-      onReadSyncIssueArticleClick={action('onReadSyncIssueArticleClick')}
-      onReadConnectivityIssueArticleClick={action(
-        'onReadConnectivityIssueArticleClick'
-      )}
-      onCheckTheTimeAgain={action('onCheckTheTimeAgain')}
-      onContinueWithoutClockSyncCheck={action(
-        'onContinueWithoutClockSyncCheck'
-      )}
-      onDownloadLogs={action('onDownloadLogs')}
-      onGetAvailableVersions={action('onGetAvailableVersions')}
-      disableDownloadLogs={boolean('disableDownloadLogs', true)}
-    />
-  ))
-  .add('Connectivity Issues', () => (
-    <Loading
-      onStatusIconClick={linkTo('Diagnostics', () => 'default')}
-      currencyIcon={currencyIcon}
-      apiIcon={apiIcon}
-      cardanoNodeState={CardanoNodeStates.UNRECOVERABLE}
-      availableAppVersion={null}
-      hasBeenConnected={boolean('hasBeenConnected', true)}
-      forceConnectivityIssue
-      isConnected={false}
-      isSynced={boolean('isSynced', true)}
-      isNodeStopping={boolean('isNodeStopping', false)}
-      isNodeStopped={boolean('isNodeStopped', false)}
-      isNotEnoughDiskSpace={boolean('isNotEnoughDiskSpace', false)}
-      isTlsCertInvalid={boolean('isTlsCertInvalid', false)}
-      diskSpaceRequired={`${text('diskSpaceRequired (GB)', 4)} GB`}
-      diskSpaceMissing={`${text('diskSpaceRequired (GB)', 1)} GB`}
-      diskSpaceRecommended={`${text('diskSpaceRequired (GB)', 8)} GB`}
-      syncPercentage={number('syncPercentage', 100)}
-      loadingDataForNextScreenMessage={{
-        id: 'loading.screen.loadingWalletData',
-        defaultMessage: '!!!Loading wallet data',
-        description: 'Message "Loading wallet data" on the loading screen.',
-      }}
-      hasLoadedCurrentLocale={boolean('hasLoadedCurrentLocale', true)}
-      hasLoadedCurrentTheme={boolean('hasLoadedCurrentTheme', true)}
-      localTimeDifference={number('localTimeDifference', 0)}
-      isSystemTimeCorrect={boolean('isSystemTimeCorrect', true)}
-      isCheckingSystemTime={boolean('isCheckingSystemTime', false)}
-      isNodeResponding={boolean('isNodeResponding', true)}
-      isNodeSubscribed={boolean('isNodeSubscribed', true)}
-      isNodeSyncing={boolean('isNodeSyncing', true)}
-      isNodeTimeCorrect={boolean('isNodeTimeCorrect', true)}
-      currentLocale="en-US"
-      currentAppVersion="1.1.1"
-      isNewAppVersionAvailable={false}
-      isNewAppVersionLoading={false}
-      isNewAppVersionLoaded
-      onExternalLinkClick={action('onExternalLinkClick')}
-      onReportIssueClick={action('onReportIssueClick')}
-      onReadSyncIssueArticleClick={action('onReadSyncIssueArticleClick')}
-      onReadConnectivityIssueArticleClick={action(
-        'onReadConnectivityIssueArticleClick'
-      )}
-      onCheckTheTimeAgain={action('onCheckTheTimeAgain')}
-      onContinueWithoutClockSyncCheck={action(
-        'onContinueWithoutClockSyncCheck'
-      )}
-      onDownloadLogs={action('onDownloadLogs')}
-      onGetAvailableVersions={action('onGetAvailableVersions')}
-      disableDownloadLogs={boolean('disableDownloadLogs', true)}
-    />
-  ))
-  .add('Sync Issues', () => (
-    <Loading
-      onStatusIconClick={linkTo('Diagnostics', () => 'default')}
-      currencyIcon={currencyIcon}
-      apiIcon={apiIcon}
-      cardanoNodeState={CardanoNodeStates.UNRECOVERABLE}
-      availableAppVersion={null}
-      hasBeenConnected={boolean('hasBeenConnected', true)}
-      forceSyncIssue
-      isConnected
-      isSynced={false}
-      isNodeStopping={boolean('isNodeStopping', false)}
-      isNodeStopped={boolean('isNodeStopped', false)}
-      isNotEnoughDiskSpace={boolean('isNotEnoughDiskSpace', false)}
-      isTlsCertInvalid={boolean('isTlsCertInvalid', false)}
-      diskSpaceRequired={`${text('diskSpaceRequired (GB)', 4)} GB`}
-      diskSpaceMissing={`${text('diskSpaceRequired (GB)', 1)} GB`}
-      diskSpaceRecommended={`${text('diskSpaceRequired (GB)', 8)} GB`}
-      syncPercentage={number('syncPercentage', 100)}
-      loadingDataForNextScreenMessage={{
-        id: 'loading.screen.loadingWalletData',
-        defaultMessage: '!!!Loading wallet data',
-        description: 'Message "Loading wallet data" on the loading screen.',
-      }}
-      hasLoadedCurrentLocale={boolean('hasLoadedCurrentLocale', true)}
-      hasLoadedCurrentTheme={boolean('hasLoadedCurrentTheme', true)}
-      localTimeDifference={number('localTimeDifference', 0)}
-      isSystemTimeCorrect={boolean('isSystemTimeCorrect', true)}
-      isCheckingSystemTime={boolean('isCheckingSystemTime', false)}
-      isNodeResponding={boolean('isNodeResponding', true)}
-      isNodeSubscribed={boolean('isNodeSubscribed', true)}
-      isNodeSyncing={boolean('isNodeSyncing', true)}
-      isNodeTimeCorrect={boolean('isNodeTimeCorrect', true)}
-      currentLocale="en-US"
-      currentAppVersion="1.1.1"
-      isNewAppVersionAvailable={false}
-      isNewAppVersionLoading={false}
-      isNewAppVersionLoaded
-      onExternalLinkClick={action('onExternalLinkClick')}
-      onReportIssueClick={action('onReportIssueClick')}
-      onReadSyncIssueArticleClick={action('onReadSyncIssueArticleClick')}
-      onReadConnectivityIssueArticleClick={action(
-        'onReadConnectivityIssueArticleClick'
-      )}
-      onCheckTheTimeAgain={action('onCheckTheTimeAgain')}
-      onContinueWithoutClockSyncCheck={action(
-        'onContinueWithoutClockSyncCheck'
-      )}
-      onDownloadLogs={action('onDownloadLogs')}
-      onGetAvailableVersions={action('onGetAvailableVersions')}
-      disableDownloadLogs={boolean('disableDownloadLogs', true)}
-    />
-  ));
+  .add('SyncingConnecting - Default', DefaultSyncingConnectingStory)
+  .add(
+    'SyncingConnecting - Connectivity Issues',
+    ConnectivityIssuesSyncingConnectingStory
+  )
+  .add('SyncingConnecting - Sync Issues', SyncIssuesSyncingConnectingStory)
+  .add('NoDiskSpaceError', NoDiskSpaceErrorStory)
+  .add('SystemTimeError', SystemTimeErrorStory)
+  .add('ManualUpdate', ManualUpdateStory);
