+++ storybook/stories/index.js
@@ -11,7 +11,6 @@ import './DeleteWalletConfirmationDialog.stories';
 import './WalletTransactionsList.stories';
 import './ExportWalletToFileDialog.stories';
 import './PaperWallets.stories';
-import './ManualUpdateOverlay.stories';
 
 // Staking
 import './Staking.stories';
